+++ resources/download/basic.ps1
@@ -7,7 +7,7 @@
 #
 
 # https://www.7-zip.org/download.html - 7-Zip - installed during start
-Get-FileFromUri -uri "https://www.7-zip.org/a/7z2301-x64.msi" -FilePath ".\downloads\7zip.msi" -CheckURL "Yes"
+$status = Get-FileFromUri -uri "https://www.7-zip.org/a/7z2301-x64.msi" -FilePath ".\downloads\7zip.msi" -CheckURL "Yes"
 
 #
 # Packages used in freshclam sandbox
@@ -15,7 +15,7 @@ Get-FileFromUri -uri "https://www.7-zip.org/a/7z2301-x64.msi" -FilePath ".\downl
 
 if ($all -or $Freshclam) {
     # ClamAV - installed during start
-    Get-GitHubRelease -repo "Cisco-Talos/clamav" -path "${SETUP_PATH}\clamav.msi" -match "win.x64.msi"
+    $status = Get-GitHubRelease -repo "Cisco-Talos/clamav" -path "${SETUP_PATH}\clamav.msi" -match "win.x64.msi"
 }
 
 #
@@ -26,7 +26,7 @@ if ($all -or $Node) {
     $nodejs = Get-DownloadUrlFromPage -url "https://nodejs.org/en/download/" -RegEx 'https:[^"]+win-x64.zip'
 
     # nodejs - installed via sandbox during download and setup of tools for dfirws
-    Get-FileFromUri -uri "${nodejs}" -FilePath ".\downloads\nodejs.zip" -CheckURL "Yes"
+    $status = Get-FileFromUri -uri "${nodejs}" -FilePath ".\downloads\nodejs.zip" -CheckURL "Yes"
 }
 
 #
@@ -44,27 +44,36 @@ if ($all -or $Python) {
     Clear-Tmp winget
 
     # Get Amazon Corretto - installed during start
-    Get-FileFromUri -uri "https://corretto.aws/downloads/latest/amazon-corretto-21-x64-windows-jdk.msi" -FilePath ".\downloads\corretto.msi"
+    $status = Get-FileFromUri -uri "https://corretto.aws/downloads/latest/amazon-corretto-21-x64-windows-jdk.msi" -FilePath ".\downloads\corretto.msi"
 
     # Ghidra - latest release
-    Get-GitHubRelease -repo "NationalSecurityAgency/ghidra" -path "${SETUP_PATH}\ghidra.zip" -match "ghidra"
-    & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa "${SETUP_PATH}\ghidra.zip" -o"${TOOLS}" | Out-Null
-    if (Test-Path "${TOOLS}\ghidra") {
-        Remove-Item "${TOOLS}\ghidra" -Recurse -Force
+    $status = Get-GitHubRelease -repo "NationalSecurityAgency/ghidra" -path "${SETUP_PATH}\ghidra.zip" -match "ghidra"
+    if ($status) {
+        & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa "${SETUP_PATH}\ghidra.zip" -o"${TOOLS}" | Out-Null
+        if (Test-Path "${TOOLS}\ghidra") {
+            Remove-Item "${TOOLS}\ghidra" -Recurse -Force
+        }
+        New-Item -ItemType Directory -Force -Path "${TOOLS}\ghidra" | Out-Null
+        Move-Item ${TOOLS}\ghidra_* "${TOOLS}\ghidra\"
     }
-    New-Item -ItemType Directory -Force -Path "${TOOLS}\ghidra" | Out-Null
-    Move-Item ${TOOLS}\ghidra_* "${TOOLS}\ghidra\"
 
     # Ghidra - older release
-    Get-FileFromUri -uri "https://github.com/NationalSecurityAgency/ghidra/releases/download/Ghidra_10.4_build/ghidra_10.4_PUBLIC_20230928.zip" -FilePath "${SETUP_PATH}\ghidra_10.4_PUBLIC_20230928.zip" -CheckURL "Yes"
-    & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa "${SETUP_PATH}\ghidra_10.4_PUBLIC_20230928.zip" -o"${TOOLS}\ghidra" | Out-Null
+    $status = Get-FileFromUri -uri "https://github.com/NationalSecurityAgency/ghidra/releases/download/Ghidra_10.4_build/ghidra_10.4_PUBLIC_20230928.zip" -FilePath "${SETUP_PATH}\ghidra_10.4_PUBLIC_20230928.zip" -CheckURL "Yes"
+    if ($status) {
+        & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa "${SETUP_PATH}\ghidra_10.4_PUBLIC_20230928.zip" -o"${TOOLS}\ghidra" | Out-Null
+    }
 
     # Ghidrathon source
-    Get-GitHubRelease -repo "mandiant/Ghidrathon" -path "${SETUP_PATH}\ghidrathon.zip" -match "Source"
-    & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa "${SETUP_PATH}\ghidrathon.zip" -o"${TOOLS}\ghidrathon" | Out-Null
+    $status = Get-GitHubRelease -repo "mandiant/Ghidrathon" -path "${SETUP_PATH}\ghidrathon.zip" -match "Source"
+    if ($status) {
+        if (Test-Path "${TOOLS}\ghidrathon") {
+            Remove-Item "${TOOLS}\ghidrathon" -Recurse -Force
+        }
+        & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa "${SETUP_PATH}\ghidrathon.zip" -o"${TOOLS}\ghidrathon" | Out-Null
+    }
 
     # Tools to compile and build - version 2019
-    Get-FileFromUri -uri "https://aka.ms/vs/16/release/vs_BuildTools.exe" -FilePath ".\downloads\vs_BuildTools.exe"
+    $status = Get-FileFromUri -uri "https://aka.ms/vs/16/release/vs_BuildTools.exe" -FilePath ".\downloads\vs_BuildTools.exe"
 }
 
 #
@@ -73,12 +82,14 @@ if ($all -or $Python) {
 
 if ($all -or $Bash) {
     # zsdt
-    Get-GitHubRelease -repo "facebook/zstd" -path "${SETUP_PATH}\zstd.zip" -match "win64.zip"
-    & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa ".\downloads\zstd.zip" -o"${TOOLS}" | Out-Null
-    if (Test-Path "${TOOLS}\zstd") {
-        Remove-Item "${TOOLS}\zstd" -Recurse -Force
+    $status = Get-GitHubRelease -repo "facebook/zstd" -path "${SETUP_PATH}\zstd.zip" -match "win64.zip"
+    if ($status) {
+        & "$env:ProgramFiles\7-Zip\7z.exe" x -aoa ".\downloads\zstd.zip" -o"${TOOLS}" | Out-Null
+        if (Test-Path "${TOOLS}\zstd") {
+            Remove-Item "${TOOLS}\zstd" -Recurse -Force
+        }
+        Move-Item ${TOOLS}\zstd-* "${TOOLS}\zstd" | Out-Null
     }
-    Move-Item ${TOOLS}\zstd-* "${TOOLS}\zstd" | Out-Null
 }
 
 #
@@ -87,7 +98,7 @@ if ($all -or $Bash) {
 
 if ($all -or $Rust) {
     # git - installed during start
-    Get-GitHubRelease -repo "git-for-windows/git" -path "${SETUP_PATH}\git.exe" -match "64-bit.exe"
+    $status = Get-GitHubRelease -repo "git-for-windows/git" -path "${SETUP_PATH}\git.exe" -match "64-bit.exe"
 
     # Rust - available for installation via dfirws-install.ps1
     Clear-Tmp winget
